ai-agent-project/requirements.txt
Pydantic==2.0.0
fastapi==0.95.0
uvicorn==0.22.0
httpx==0.23.0
pyyaml==6.0
openai==0.27.0
ollama==0.1.0
lmstudio==0.1.0
pytest==7.2.0
mypy==0.991